+++ synccommerce/src/sdk/models/operations/createconnection.ts
@@ -5,10 +5,17 @@
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import * as shared from "../shared";
 import { AxiosResponse } from "axios";
+import { Expose } from "class-transformer";
+
+export class CreateConnectionRequestBody extends SpeakeasyBase {
+    @SpeakeasyMetadata()
+    @Expose({ name: "platformKey" })
+    platformKey?: string;
+}
 
 export class CreateConnectionRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "request, media_type=application/json" })
-    requestBody?: string;
+    requestBody?: CreateConnectionRequestBody;
 
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
@@ -16,13 +23,19 @@ export class CreateConnectionRequest extends SpeakeasyBase {
 
 export class CreateConnectionResponse extends SpeakeasyBase {
     /**
-     * Success
+     * OK
      */
     @SpeakeasyMetadata()
     connection?: shared.Connection;
 
     @SpeakeasyMetadata()
     contentType: string;
+
+    /**
+     * Your API request was not properly authorized.
+     */
+    @SpeakeasyMetadata()
+    errorMessage?: shared.ErrorMessage;
 
     @SpeakeasyMetadata()
     statusCode: number;